Buzzer- The Daily Scheduler
Buzzer is a repository for End User who wishes to keep a track of the
checklist and reminders daily. Tools Used: JavaFX and SQlite for
Database. Stakeholders and Interest: End User who wishes to keep a track
of their work and the time spent on various categories. Preconditions:
1. User has to input the checks who wish to complete on that day. 2.
Additionally, can add reminders along with the time on which he wants to
get reminded. Success Guarantee – Post Conditions: 1. BUZZER start page
is shown. User has to input his user login and password. 2. BUZZER main
page is shown which has 3 tabs- Checks, Reminders, Tasks. 3. User can
input the checks before starting his day. 4. User can input the
reminders in the Reminder tab along with date and time. Alarm is set for
that date and time. 5. Whenever a check is completed, user can mark that
check as Done. 6. When the clock hits the reminder time, it alarms out
along with the respective reminder. 7. User can delete the check if he
wishes not be stored in the database. 8. User can view the columns as
per his choice by clicking on “+” of the table. 9. User can sort the
table as per his choice. For eg: if he wishes to see the urgent task in
his list, he can sort and check. 10. When clicked on Report or Brief
from the MenuBar, it gives the Total Done and Undone task in a table
along with the PIE Chart of the Task spent with respect to categories.
11. Before User wants to exit the application, it is prompted with an
alert box stating that he has certain number of checks and reminders
undone. 12. Greeting messages are shows as per the Time of the day.